Change log handler to logf

+++ registry/registry.go
@@ -21,7 +21,7 @@ func Quiet(format string, args ...interface{}) {
  * Pass log messages along to Go's "log" module.
  */
 func Log(format string, args ...interface{}) {
-	log.Printf(format, args...)
+	logf.Printf(format, args...)
 }
 
 type Registry struct {
